requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0 
pytest==7.4.0
pylint==2.17.5
mypy==1.5.1
coverage==7.3.1
jsonschema==4.19.0
GitPython==3.1.32
requests==2.28.1
fastapi==0.115.4
uvicorn==0.25.0
pydantic==2.5.2
aiofiles==23.2.1
cryptography>=3.4.8
